Disable interleave-pdf-mode along with interleave
When `interleave` is disabled in the org notes buffer, also disable in
the `interleave-pdf-mode` in the PDF file too.

Now if the PDF file buffer is already killed by the user, and if the
interleave-pdf-buffer stored the buffer object instead of buffer name,
`(get-buffer interleave-pdf-buffer)` will still return a non-nil value.

But if `interleave-pdf-buffer` stores the buffer **name** instead of the
object handle, `(get-buffer interleave-pdf-buffer)` will return nil if
user already killed the PDF file buffer.

For that reason and for consistency, now both interleave-pdf-buffer and
interleave-org-buffer store the buffer name instead of object.

Along with this change, following the naming convention of elisp
variables,

- `*interleave--pdf-buffer*` is renamed to `interleave-pdf-buffer`
- `*interleave--org-buffer*` is renamed to `interleave-org-buffer`