Better mode enabling and disabling handles.
Currently when there's no interleave property found
and require user to input some path, if a C-g or something
else terminate the progress, the `interleave-mode` is
still enabled, and the window config is changed, namely,
beingg split. This commit will handle this situaion. If
the pdf file is not successfully input or find. `interleave-mode`
will be disabled automatically, while preserving the origin
window config.

Also, when `interleave-mode` is disabled, the respective
variables are finialized.